requirements.txt
uvicorn~=0.35.0
python-dotenv~=1.1.1
fastapi~=0.115.14
SQLAlchemy~=2.0.41
pytesseract~=0.3.13
pillow~=11.2.1
celery~=5.5.3
redis==5.0.0
python-multipart==0.0.19